+++ src/Commands/Utility/MDN.ts
@@ -1,33 +1,10 @@
 import { Command } from "../../Structures/Command";
 import { Message } from "discord.js";
-import { get } from 'https';
+import { mdn } from "../../Backend/Commands/MDNHandler";
 import Embed from "../../Structures/Embed";
+import { compareTwoStrings } from "../../Backend/Helpers/CompareStrings";
 
 type mdnResult = { partialURL: string, title: string, diff: number }
-
-/**
- * Compare similarity between 2 strings using Dice's coefficient. 
- * Case sensitive - ``abc`` isn't ``AbC`` in my opinion.
- * @see https://en.wikipedia.org/wiki/S%C3%B8rensen%E2%80%93Dice_coefficient
- * @author Khafra
- * @param X first string
- * @param Y second string
- * @returns {number} "quotient of similarity" (number 0-1).
- */
-const compareTwoStrings = (X: string, Y: string): number => {
-    const bigramsX = Array.from(
-        Array(X.length - 1),
-        (_, index) => X[index] + X[index + 1]
-    );
-
-    const bigramsY = Array.from(
-        Array(Y.length - 1),
-        (_, index) => Y[index] + Y[index + 1]
-    );
-
-    const inBoth = bigramsX.filter(current => bigramsY.indexOf(current) > -1);
-    return (2 * inBoth.length) / (bigramsX.length + bigramsY.length);
-}
 
 export default class extends Command {
     constructor() {
@@ -51,7 +28,7 @@ export default class extends Command {
 
         let parsed: mdnResult[] = [];
         try {
-            const html = await this.mdn(args.join(' '));
+            const html = await mdn(args.join(' '));
             parsed = this.parseHTML(html, args.join(' '));
         } catch {
             return message.channel.send(Embed.fail('An unexpected error occurred!'));
@@ -64,26 +41,6 @@ export default class extends Command {
             .setTimestamp()
 
         return message.channel.send(embed);
-    }
-
-    /**
-     * Get the search result page's HTML
-     * @param q query
-     */
-    mdn(q: string): Promise<any> {
-        q = encodeURIComponent(q.replace(' ', '+'))
-
-        return new Promise((resolve, reject) => {
-            get('https://developer.mozilla.org/en-US/search?q=' + q, res => {
-                if(res.statusCode !== 200) {
-                    return reject(new Error('Non-200 status code'));
-                }
-    
-                const body = [];
-                res.on('data', d => body.push(d));
-                res.on('end', () => resolve(body.join(' ')));
-            }).on('error', e => reject(e));
-        });
     }
 
     /**