+++ src/Events/Message.ts
@@ -1,8 +1,8 @@
 import { Event } from "../Structures/Event";
-import { Message, Permissions } from "discord.js";
-import { dbHelpers } from "../Helpers/GuildSettings";
+import { Message, Permissions, PermissionString } from "discord.js";
+import { dbHelpers } from "../Backend/Helpers/GuildSettings";
 import KhafraClient from "../Bot/KhafraBot";
-import { Sanitize } from "../Helpers/SanitizeCommand";
+import { Sanitize } from "../Backend/Helpers/SanitizeCommand";
 import Cooldown from "../Structures/Cooldown";
 import Embed from "../Structures/Embed";
 
@@ -24,6 +24,30 @@ export default class extends Event {
         const prefix = settings?.prefix ?? '!';
         const command = KhafraClient.Commands.get(cmd.slice(prefix.length));
 
+        if(settings?.reacts) {
+            const perms = message.guild.me.permissionsIn(message.channel);
+            const hasPerms = [
+                'READ_MESSAGE_HISTORY',
+                'ADD_REACTIONS'
+            ] as PermissionString[];
+            
+            if(!hasPerms.every(perm => perms.has(perm))) {
+                return;
+            }
+    
+            const user = settings.reacts.filter(r => r.id === message.author.id).pop();
+            if(user) {
+                const random = Math.floor(Math.random() * 100 + 1);
+                const chance = random <= +user.chance;
+
+                if(chance) {
+                    try {
+                        message.react(user.emoji.replace(/\\/g, '')); // SQLite sanitizes input, in this case by adding backslashes.
+                    } catch {} // doesn't really matter if this fails.
+                }
+            }
+        }
+
         if(!command) {
             return;
         } else if(Cooldown.$has(message.author.id, command.name)) {
@@ -32,25 +56,6 @@ export default class extends Event {
             `));
         } else {
             Cooldown.$set(message.author.id, command.name, command.cooldown);
-        }
-    
-        if(settings?.reacts) {
-            const botPerms = message.guild.me.permissions;
-            const hasPerms = [ Permissions.FLAGS.ADD_REACTIONS, Permissions.FLAGS.READ_MESSAGE_HISTORY ].every(p => botPerms.has(p));
-            if(!hasPerms) {
-                return;
-            }
-    
-            const user = settings.reacts.filter((r: any) => r.id === message.author.id).pop();
-    
-            if(user) {
-                const chance = Math.floor(Math.random() * 100 + 1) <= +user.chance;
-                if(chance) {
-                    try {
-                        message.react(user.emoji.replace(/\\/g, '')); // SQLite sanitizes input, in this case by adding backslashes.
-                    } catch {} // doesn't really matter if this fails.
-                }
-            }
         }
     
         if(!cmd.startsWith(prefix)) {