+++ src/Events/MessageReactionRemove.ts
@@ -1,6 +1,6 @@
 import { Event } from "../Structures/Event";
-import { MessageReaction, User, PartialUser } from "discord.js";
-import { dbHelpers, react_messages } from "../Helpers/GuildSettings";
+import { MessageReaction, User, PartialUser, PermissionString } from "discord.js";
+import { dbHelpers } from "../Backend/Helpers/GuildSettings";
 
 export default class extends Event {
     constructor() {
@@ -12,11 +12,23 @@ export default class extends Event {
             await reaction.fetch();
         }
     
+        if(reaction.message.deleted) {
+            return;
+        }
+    
         if(user.partial) {
             await user.fetch();
         }
     
+        const perms = reaction.message.guild.me.permissionsIn(reaction.message.channel);
+        const needed = [
+            'READ_MESSAGE_HISTORY',
+            'MANAGE_ROLES' // not sure if required
+        ] as PermissionString[];
+        
         if(user.id === reaction.message.client.user.id) {
+            return;
+        } else if(!needed.every(perm => perms.has(perm))) {
             return;
         }
     
@@ -25,11 +37,11 @@ export default class extends Event {
             return;
         }
     
-        const filtered: any[] = guildSettings.react_messages.filter((r: react_messages) => {
+        const filtered = guildSettings.react_messages.filter(r => {
             const emoji = (reaction.message.client.emojis.resolve(r.emoji) ?? r.emoji) as any;
             if(
-                (emoji === r.emoji || emoji?.id === r.emoji) &&
-                r.id === reaction.message.id
+                (emoji === r.emoji || emoji?.id === r.emoji) && // Emoji is the same
+                r.id === reaction.message.id                    // message id is the same
             ) {
                 return r;
             }
@@ -39,9 +51,10 @@ export default class extends Event {
             return;
         }
     
+        // member MUST be fetched or they will never be manageable!
         const member = await reaction.message.guild.members.fetch(user.id);
-        if(member.manageable) {
-            return member.roles.remove(filtered[0].role, 'Un-reacted'); // yeah, this is the only line that was edited from MessageReactionAdd
+        if(member.manageable) {    
+            return member.roles.remove(filtered[0].role, 'Reacted');
         }
     }
 }